+++ WebJobs.Extensions.Redis.sln
@@ -11,13 +11,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.Functions.W
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.WebJobs.Extensions.Redis", "src\Microsoft.Azure.WebJobs.Extensions.Redis\Microsoft.Azure.WebJobs.Extensions.Redis.csproj", "{A1A50657-EB50-4B7C-9BDD-AB2A357F8235}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReadThroughSamples", "samples\CosmosDbIntegtation\ReadThroughSamples\ReadThroughSamples.csproj", "{04D19EE9-953A-4D5D-99D5-345692237B35}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReadThroughSamples", "samples\CosmosDBIntegration\ReadThroughSamples\ReadThroughSamples.csproj", "{77A0D8F6-CAB3-450B-BBCB-658C8A47EB3C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WriteAroundSamples", "samples\CosmosDbIntegtation\WriteAroundSamples\WriteAroundSamples.csproj", "{D47E35C1-6BE9-4858-BB3A-806EB6910A07}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WriteAroundSamples", "samples\CosmosDBIntegration\WriteAroundSamples\WriteAroundSamples.csproj", "{780469B0-E554-46F8-8333-4C1630C04E3D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WriteBehindSamples", "samples\CosmosDbIntegtation\WriteBehindSamples\WriteBehindSamples.csproj", "{C12535DC-5EBF-4037-B8DE-FDC36353DE7A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WriteBehindSamples", "samples\CosmosDBIntegration\WriteBehindSamples\WriteBehindSamples.csproj", "{C6273150-22D0-4967-B7B6-9A2EA674053D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WriteThroughSamples", "samples\CosmosDbIntegtation\WriteThroughSamples\WriteThroughSamples.csproj", "{93846DFB-E487-4908-B343-CEC83D25E762}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WriteThroughSamples", "samples\CosmosDBIntegration\WriteThroughSamples\WriteThroughSamples.csproj", "{61001020-ABEE-4B06-BF56-1C1989838CBA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -41,22 +41,22 @@ Global
 		{A1A50657-EB50-4B7C-9BDD-AB2A357F8235}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A1A50657-EB50-4B7C-9BDD-AB2A357F8235}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A1A50657-EB50-4B7C-9BDD-AB2A357F8235}.Release|Any CPU.Build.0 = Release|Any CPU
-		{04D19EE9-953A-4D5D-99D5-345692237B35}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{04D19EE9-953A-4D5D-99D5-345692237B35}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{04D19EE9-953A-4D5D-99D5-345692237B35}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{04D19EE9-953A-4D5D-99D5-345692237B35}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D47E35C1-6BE9-4858-BB3A-806EB6910A07}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D47E35C1-6BE9-4858-BB3A-806EB6910A07}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D47E35C1-6BE9-4858-BB3A-806EB6910A07}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D47E35C1-6BE9-4858-BB3A-806EB6910A07}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C12535DC-5EBF-4037-B8DE-FDC36353DE7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C12535DC-5EBF-4037-B8DE-FDC36353DE7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C12535DC-5EBF-4037-B8DE-FDC36353DE7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C12535DC-5EBF-4037-B8DE-FDC36353DE7A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{93846DFB-E487-4908-B343-CEC83D25E762}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{93846DFB-E487-4908-B343-CEC83D25E762}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{93846DFB-E487-4908-B343-CEC83D25E762}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{93846DFB-E487-4908-B343-CEC83D25E762}.Release|Any CPU.Build.0 = Release|Any CPU
+		{77A0D8F6-CAB3-450B-BBCB-658C8A47EB3C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{77A0D8F6-CAB3-450B-BBCB-658C8A47EB3C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{77A0D8F6-CAB3-450B-BBCB-658C8A47EB3C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{77A0D8F6-CAB3-450B-BBCB-658C8A47EB3C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{780469B0-E554-46F8-8333-4C1630C04E3D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{780469B0-E554-46F8-8333-4C1630C04E3D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{780469B0-E554-46F8-8333-4C1630C04E3D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{780469B0-E554-46F8-8333-4C1630C04E3D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C6273150-22D0-4967-B7B6-9A2EA674053D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6273150-22D0-4967-B7B6-9A2EA674053D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6273150-22D0-4967-B7B6-9A2EA674053D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6273150-22D0-4967-B7B6-9A2EA674053D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{61001020-ABEE-4B06-BF56-1C1989838CBA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{61001020-ABEE-4B06-BF56-1C1989838CBA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{61001020-ABEE-4B06-BF56-1C1989838CBA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{61001020-ABEE-4B06-BF56-1C1989838CBA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
