+++ samples/CosmosDBIntegration/Models/RedisData.cs
