+++ samples/CosmosDBIntegration/ReadThroughSamples/Models/RedisData.cs
