+++ samples/CosmosDBIntegration/ReadThroughSamples/Properties/serviceDependencies.local.json
