+++ samples/CosmosDBIntegration/ReadThroughSamples/ReadThroughSamples.csproj
