+++ samples/CosmosDBIntegration/WriteAroundSamples/ListSample.cs
