+++ samples/CosmosDBIntegration/WriteAroundSamples/Models/RedisData.cs
