+++ samples/CosmosDBIntegration/WriteAroundSamples/Properties/serviceDependencies.json
