+++ samples/CosmosDBIntegration/WriteBehindSamples/Models/RedisData.cs
