+++ samples/CosmosDBIntegration/WriteBehindSamples/Properties/launchSettings.json
