+++ samples/CosmosDBIntegration/WriteBehindSamples/Properties/serviceDependencies.json
