+++ samples/CosmosDBIntegration/WriteBehindSamples/PubSubSample.cs
