+++ samples/CosmosDBIntegration/WriteThroughSamples/Properties/launchSettings.json
