+++ samples/CosmosDBIntegration/WriteThroughSamples/host.json
