--- samples/CosmosDbIntegtation/WriteThroughSamples/local.settings.json
@@ -1,15 +1,0 @@
-{
-  "IsEncrypted": false,
-  "Values": {
-    "AzureWebJobsStorage": "UseDevelopmentStorage=true",
-    "Values": {
-      "AzureWebJobsStorage": "UseDevelopmentStorage=true",
-      "FUNCTIONS_WORKER_RUNTIME": "dotnet",
-      "CosmosDbDatabaseId": "DatabaseId",
-      "CosmosDbContainerId": "ContainerId",
-      "PubSubContainerId": "PSContainerId",
-      "PubSubChannel": "PubSubTest",
-      "RedisConnectionString": "<cache-name>.redis.cache.windows.net:6380,password=<access-key>,ssl=True,abortConnect=False,tiebreaker=",
-      "CosmosDbConnectionString": "AccountEndpoint=https://<cosmosdb-account>.documents.azure.com:443/;AccountKey=<access-key>;"
-    }
-  }
